Seperated Parsing and Output from Workflows
Seperated the Output and Parsing from the DGeoFlow work flow.
