+++ applications/GeoMechanicsApplication/custom_workflows/dgeoflow.cpp
@@ -20,6 +20,8 @@
 #include "input_output/logger.h"
 #include "input_output/logger_output.h"
 #include "input_output/logger_table_output.h"
+#include "dgeooutput.h"
+#include "dgeoparser.h"
 
 class GeoFlowApplyConstantScalarValueProcess : public Kratos::ApplyConstantScalarValueProcess
 {
@@ -72,44 +74,6 @@ public:
         return mVariableName == "WATER_PRESSURE";
     }
 };
-
-void NodeOperation::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part){};
-
-void NodeDISPLACEMENT::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.WriteNodalResults(Kratos::DISPLACEMENT, model_part.Nodes(), 0, 0); }
-
-void NodeTOTAL_DISPLACEMENT::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.WriteNodalResults(Kratos::TOTAL_DISPLACEMENT, model_part.Nodes(), 0, 0); }
-
-void NodeWATER_PRESSURE::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.WriteNodalResults(Kratos::WATER_PRESSURE, model_part.Nodes(), 0, 0); }
-
-void NodeNORMAL_FLUID_FLUX::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.WriteNodalResults(Kratos::NORMAL_FLUID_FLUX, model_part.Nodes(), 0, 0); }
-
-void NodeVOLUME_ACCELERATION::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.WriteNodalResults(Kratos::VOLUME_ACCELERATION, model_part.Nodes(), 0, 0); }
-
-void NodeHYDRAULIC_DISCHARGE::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.WriteNodalResults(Kratos::HYDRAULIC_DISCHARGE, model_part.Nodes(), 0, 0); }
-
-void NodeHYDRAULIC_HEAD::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.WriteNodalResults(Kratos::HYDRAULIC_HEAD, model_part.Nodes(), 0, 0); }
-
-void GaussOperation::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part){};
-
-void GaussFLUID_FLUX_VECTOR::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::FLUID_FLUX_VECTOR, model_part, 0, 0); }
-
-void GaussHYDRAULIC_HEAD::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::HYDRAULIC_HEAD, model_part, 0, 0); }
-
-void GaussLOCAL_FLUID_FLUX_VECTOR::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::LOCAL_FLUID_FLUX_VECTOR, model_part, 0, 0); }
-
-void GaussLOCAL_PERMEABILITY_MATRIX::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::LOCAL_PERMEABILITY_MATRIX, model_part, 0, 0); }
-
-void GaussPERMEABILITY_MATRIX::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::PERMEABILITY_MATRIX, model_part, 0, 0); }
-
-void GaussDEGREE_OF_SATURATION::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::DEGREE_OF_SATURATION, model_part, 0, 0); }
-
-void GaussDERIVATIVE_OF_SATURATION::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::DERIVATIVE_OF_SATURATION, model_part, 0, 0); }
-
-void GaussRELATIVE_PERMEABILITY::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::RELATIVE_PERMEABILITY, model_part, 0, 0); }
-
-void GaussPIPE_ACTIVE::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::PIPE_ACTIVE, model_part, 0, 0); }
-
-void GaussPIPE_HEIGHT::write(Kratos::GidIO<> &gid_io, Kratos::ModelPart &model_part) { gid_io.PrintOnGaussPoints(Kratos::PIPE_HEIGHT, model_part, 0, 0); }
 
 namespace Kratos
 {
@@ -220,205 +184,6 @@ namespace Kratos
         return p_solving_strategy;
     }
 
-    void KratosExecute::parseMesh(ModelPart &model_part, std::string filepath)
-    {
-        // Parses MDPA file into model_part
-        std::ifstream input(filepath);
-        bool read_properties = false;
-        bool read_nodes = false;
-        bool read_elements = false;
-
-        bool read_subparts = false;
-        bool read_subparts_table = false;
-        bool read_subparts_nodes = false;
-        bool read_subparts_elements = false;
-        bool read_subparts_conditions = false;
-
-        std::string element_type;
-        std::string part_name;
-        std::string nodeStr;
-
-        for (std::string line; getline(input, line);)
-        {
-
-            //===================== Properties =========================
-            if (line.substr(0, 16) == "Begin Properties")
-            {
-                read_properties = true;
-                std::size_t found = line.find_last_of(" ");
-                int property_id = stoi(line.substr(found + 1));
-                model_part.CreateNewProperties(property_id);
-                continue;
-            }
-            if (line == "End Properties")
-            {
-                read_properties = false;
-                continue;
-            }
-            if (read_properties)
-            {
-                KRATOS_ERROR << "Reading Properties - Not Implemented " << std::endl;
-            }
-            //=====================   Nodes   =========================
-            if (line == "Begin Nodes")
-            {
-                read_nodes = true;
-                continue;
-            }
-            if (line == "End Nodes")
-            {
-                read_nodes = false;
-                continue;
-            }
-            if (read_nodes)
-            {
-                std::istringstream iss(line);
-                int nodeId;
-                double x, y, z;
-                iss >> nodeId >> x >> y >> z;
-                model_part.CreateNewNode(nodeId, x, y, z);
-            }
-            //====================   Element   ==========================
-            if (line.substr(0, 14) == "Begin Elements")
-            {
-                read_elements = true;
-                std::size_t found = line.find_last_of(" ");
-                element_type = line.substr(found + 1);
-
-                std::size_t posD = element_type.find_last_of("D");
-                nodeStr = element_type.substr(posD + 1);
-
-                continue;
-            }
-            if (line == "End Elements")
-            {
-                read_elements = false;
-                continue;
-            }
-
-            if (read_elements)
-            {
-                unsigned long elementId, propertyId, node1, node2, node3;
-                std::vector<ModelPart::IndexType> element_nodes;
-                std::istringstream iss(line);
-
-                if (nodeStr == "4N")
-                {
-                    unsigned long node4;
-                    iss >> elementId >> propertyId >> node1 >> node2 >> node3 >> node4;
-                    element_nodes = {node1, node2, node3, node4};
-                }
-                else if (nodeStr == "3N")
-                {
-                    iss >> elementId >> propertyId >> node1 >> node2 >> node3;
-                    element_nodes = {node1, node2, node3};
-                }
-                else
-                {
-                    KRATOS_ERROR << "Element Type Unknown / Not Implemented " << std::endl;
-                }
-                auto p_elem_prop = model_part.pGetProperties(propertyId);
-                model_part.CreateNewElement(element_type, elementId, element_nodes, p_elem_prop);
-            }
-
-            //===================== Properties =========================
-
-            if (line.substr(0, 18) == "Begin SubModelPart")
-            {
-                read_subparts = true;
-                std::size_t found = line.find_last_of(" ");
-                part_name = line.substr(found + 1);
-                model_part.CreateSubModelPart(part_name);
-                continue;
-            }
-            if (line == "End SubModelPart")
-            {
-                read_subparts = false;
-                continue;
-            }
-            if (read_subparts)
-            {
-                auto subpart = model_part.pGetSubModelPart(part_name);
-                //===========  Sub-Tables  ===============
-                if (line == "  Begin SubModelPartTables")
-                {
-                    read_subparts_table = true;
-                    continue;
-                }
-                if (line == "  End SubModelPartTables")
-                {
-                    read_subparts_table = false;
-                    continue;
-                }
-                if (read_subparts_table)
-                {
-                    KRATOS_ERROR << "Subpart Tables - Not Implemented " << std::endl;
-                }
-
-                //===========  Sub-Nodes  ===============
-
-                if (line == "  Begin SubModelPartNodes")
-                {
-                    read_subparts_nodes = true;
-                    continue;
-                }
-                if (line == "  End SubModelPartNodes")
-                {
-                    read_subparts_nodes = false;
-                    continue;
-                }
-                if (read_subparts_nodes)
-                {
-                    auto node = model_part.pGetNode(stoi(line));
-                    subpart->AddNode(node);
-                }
-
-                //===========  Sub-Elements  ===============
-
-                if (line == "  Begin SubModelPartElements")
-                {
-                    read_subparts_elements = true;
-                    continue;
-                }
-                if (line == "  End SubModelPartElements")
-                {
-                    read_subparts_elements = false;
-                    continue;
-                }
-                if (read_subparts_elements)
-                {
-                    auto element = model_part.pGetElement(stoi(line));
-                    subpart->AddElement(element);
-                }
-
-                //===========  Sub-Elements  ===============
-
-                if (line == "  Begin SubModelPartConditions")
-                {
-                    read_subparts_conditions = true;
-                    continue;
-                }
-                if (line == "  End SubModelPartConditions")
-                {
-                    read_subparts_conditions = false;
-                    continue;
-                }
-                if (read_subparts_conditions)
-                {
-                    KRATOS_ERROR << "Subpart Conditions - Not Implemented " << std::endl;
-                }
-            }
-        }
-        input.close();
-    }
-
-    void KratosExecute::parseMaterial(Model &model, std::string filepath)
-    {
-        std::string parameters = "{ \"Parameters\" : { \"materials_filename\" :\"" + filepath + "\"}}";
-        Parameters material_file{parameters};
-        ReadMaterialsUtility(material_file, model);
-    }
-
     Parameters KratosExecute::openProjectParamsFile(std::string filepath)
     {
         std::ifstream t(filepath);
@@ -426,176 +191,6 @@ namespace Kratos
         buffer << t.rdbuf();
         Parameters projFile{buffer.str()};
         return projFile;
-    }
-
-    std::vector<std::shared_ptr<Process>> KratosExecute::parseProcess(ModelPart &model_part, Parameters projFile)
-    {
-        // Currently: In DGeoflow only fixed hydrostatic head has been , also need load of gravity.
-
-        std::vector<std::shared_ptr<Process>> processes;
-
-        auto constraints_processes = projFile["processes"]["constraints_process_list"];
-        for (Parameters process : constraints_processes)
-        {
-            // we only support fixed hydrostatic head
-            auto name = process["Parameters"]["model_part_name"].GetString();
-            auto pressure_type = process["Parameters"]["fluid_pressure_type"].GetString();
-
-            std::size_t found = name.find_last_of(".");
-            std::string subname = name.substr(found + 1);
-
-            ModelPart &part = model_part.GetSubModelPart(subname);
-
-            if (pressure_type == "Uniform")
-            {
-                auto value = process["Parameters"]["value"].GetDouble();
-                processes.push_back(make_shared<GeoFlowApplyConstantScalarValueProcess>(GeoFlowApplyConstantScalarValueProcess(part, WATER_PRESSURE,
-                                                                                                                               value, 0, GeoFlowApplyConstantScalarValueProcess::VARIABLE_IS_FIXED)));
-            }
-            else if (pressure_type == "Hydrostatic")
-            {
-                auto cProcesses = process.Clone();
-                cProcesses["Parameters"].RemoveValue("fluid_pressure_type");
-                processes.push_back(make_shared<GeoFlowApplyConstantHydrostaticPressureProcess>(GeoFlowApplyConstantHydrostaticPressureProcess(part, cProcesses["Parameters"])));
-            }
-            else
-            {
-                KRATOS_ERROR << "Reading Processing - Not Implemented - Pressure_type" << std::endl;
-            }
-        }
-
-        auto loads_processes = projFile["processes"]["loads_process_list"];
-        // Should only have one.
-        auto name = loads_processes.GetArrayItem(0)["Parameters"]["model_part_name"].GetString();
-        std::size_t found = name.find_last_of(".");
-        std::string subname = name.substr(found + 1);
-        ModelPart &part = model_part.GetSubModelPart(subname);
-        processes.push_back(make_shared<ApplyConstantScalarValueProcess>(ApplyConstantScalarValueProcess(part, VOLUME_ACCELERATION_X,
-                                                                                                         0.0, 0, ApplyConstantScalarValueProcess::VARIABLE_IS_FIXED)));
-
-        processes.push_back(make_shared<ApplyConstantScalarValueProcess>(ApplyConstantScalarValueProcess(part, VOLUME_ACCELERATION_Y, -9.81,
-                                                                                                         0, ApplyConstantScalarValueProcess::VARIABLE_IS_FIXED)));
-
-        processes.push_back(make_shared<Process>(ApplyConstantScalarValueProcess(part, VOLUME_ACCELERATION_Z, 0.0,
-                                                                                 0, ApplyConstantScalarValueProcess::VARIABLE_IS_FIXED)));
-
-        return processes;
-    }
-
-    void KratosExecute::outputGiD(Model &model, ModelPart &model_part, Parameters parameters, std::string workingDirectory)
-    {
-        std::map<std::string, GiD_PostMode> PostMode;
-        PostMode["GiD_PostAscii"] = GiD_PostAscii;
-        PostMode["GiD_PostAsciiZipped"] = GiD_PostAsciiZipped;
-        PostMode["GiD_PostBinary"] = GiD_PostBinary;
-        PostMode["GiD_PostHDF5"] = GiD_PostHDF5;
-
-        std::map<std::string, MultiFileFlag> MultiFiles;
-        MultiFiles["SingleFile"] = SingleFile;
-        MultiFiles["MultipleFiles"] = MultipleFiles;
-
-        std::map<std::string, WriteDeformedMeshFlag> DeformedFlag;
-        DeformedFlag["WriteDeformed"] = WriteDeformed;
-        DeformedFlag["WriteUndeformed"] = WriteUndeformed;
-
-        std::map<std::string, WriteConditionsFlag> ConditionFlag;
-        ConditionFlag["WriteConditions"] = WriteConditions;
-        ConditionFlag["WriteElementsOnly"] = WriteElementsOnly;
-        ConditionFlag["WriteConditionsOnly"] = WriteConditionsOnly;
-
-        Parameters gid_out = parameters["output_processes"]["gid_output"].GetArrayItem(0);
-        Parameters outputParameters = gid_out["Parameters"];
-        std::string filename = outputParameters["output_name"].GetString();
-        GiD_PostMode gid_output_type = PostMode[outputParameters["postprocess_parameters"]["result_file_configuration"]["gidpost_flags"]["GiDPostMode"].GetString()];
-        MultiFileFlag multifiles_output = MultiFiles[outputParameters["postprocess_parameters"]["result_file_configuration"]["gidpost_flags"]["MultiFileFlag"].GetString()];
-        WriteDeformedMeshFlag deformed_output = DeformedFlag[outputParameters["postprocess_parameters"]["result_file_configuration"]["gidpost_flags"]["WriteDeformedMeshFlag"].GetString()];
-        WriteConditionsFlag condition_output = ConditionFlag[outputParameters["postprocess_parameters"]["result_file_configuration"]["gidpost_flags"]["WriteConditionsFlag"].GetString()];
-
-        filename = workingDirectory + "/" + filename;
-        GidIO<> gid_io(filename, gid_output_type, multifiles_output, deformed_output, condition_output);
-
-        gid_io.InitializeMesh(0.0);
-        gid_io.WriteMesh(model_part.GetMesh());
-        gid_io.FinalizeMesh();
-        gid_io.InitializeResults(0, model_part.GetMesh());
-
-        std::unordered_map<std::string, unique_ptr<NodeOperation>> NodeOutput;
-        NodeOutput["DISPLACEMENT"] = Kratos::make_unique<NodeDISPLACEMENT>();
-        NodeOutput["TOTAL_DISPLACEMENT"] = Kratos::make_unique<NodeTOTAL_DISPLACEMENT>();
-        NodeOutput["WATER_PRESSURE"] = Kratos::make_unique<NodeWATER_PRESSURE>();
-        NodeOutput["NORMAL_FLUID_FLUX"] = Kratos::make_unique<NodeNORMAL_FLUID_FLUX>();
-        NodeOutput["VOLUME_ACCELERATION"] = Kratos::make_unique<NodeVOLUME_ACCELERATION>();
-        NodeOutput["HYDRAULIC_DISCHARGE"] = Kratos::make_unique<NodeHYDRAULIC_DISCHARGE>();
-        NodeOutput["HYDRAULIC_HEAD"] = Kratos::make_unique<NodeHYDRAULIC_HEAD>();
-
-        // Calculate hydraulic head on the nodes
-        auto gauss_outputs = outputParameters["postprocess_parameters"]["result_file_configuration"]["gauss_point_results"].GetStringArray();
-        auto nodal_outputs = outputParameters["postprocess_parameters"]["result_file_configuration"]["nodal_results"].GetStringArray();
-
-        if (std::find(gauss_outputs.begin(), gauss_outputs.end(), "HYDRAULIC_HEAD") != gauss_outputs.end())
-        {
-            calculateNodalHydraulicHead(gid_io, model_part);
-        }
-
-        for (std::string var : nodal_outputs)
-        {
-            NodeOutput[var]->write(gid_io, model_part);
-        }
-
-        std::unordered_map<std::string, std::unique_ptr<GaussOperation>> GaussOutput;
-        GaussOutput["FLUID_FLUX_VECTOR"] = Kratos::make_unique<GaussFLUID_FLUX_VECTOR>();
-        GaussOutput["HYDRAULIC_HEAD"] = Kratos::make_unique<GaussHYDRAULIC_HEAD>();
-        GaussOutput["LOCAL_FLUID_FLUX_VECTOR"] = Kratos::make_unique<GaussLOCAL_FLUID_FLUX_VECTOR>();
-        GaussOutput["LOCAL_PERMEABILITY_MATRIX"] = Kratos::make_unique<GaussLOCAL_PERMEABILITY_MATRIX>();
-        GaussOutput["PERMEABILITY_MATRIX"] = Kratos::make_unique<GaussPERMEABILITY_MATRIX>();
-        GaussOutput["DEGREE_OF_SATURATION"] = Kratos::make_unique<GaussDEGREE_OF_SATURATION>();
-        GaussOutput["DERIVATIVE_OF_SATURATION"] = Kratos::make_unique<GaussDERIVATIVE_OF_SATURATION>();
-        GaussOutput["RELATIVE_PERMEABILITY"] = Kratos::make_unique<GaussRELATIVE_PERMEABILITY>();
-        GaussOutput["PIPE_ACTIVE"] = Kratos::make_unique<GaussPIPE_ACTIVE>();
-        GaussOutput["PIPE_HEIGHT"] = Kratos::make_unique<GaussPIPE_HEIGHT>();
-
-        // Now Output Gauss Point Results on Gauss Points
-        for (std::string var : gauss_outputs)
-        {
-            GaussOutput[var]->write(gid_io, model_part);
-        }
-
-        gid_io.FinalizeResults();
-    }
-
-    void KratosExecute::calculateNodalHydraulicHead(GidIO<> &gid_io, ModelPart &model_part) {
-            auto element_var = &(KratosComponents<Variable<double>>::Get("HYDRAULIC_HEAD"));
-
-            for (Element element : model_part.Elements())
-            {
-                auto rGeom = element.GetGeometry();
-                auto rProp = element.GetProperties();
-                
-                for (unsigned int node = 0; node < 3; ++node)
-                {
-                    array_1d<double, 3> NodeVolumeAcceleration;
-                    noalias(NodeVolumeAcceleration) = rGeom[node].FastGetSolutionStepValue(VOLUME_ACCELERATION, 0);
-                    const double g = norm_2(NodeVolumeAcceleration);
-                    if (g > std::numeric_limits<double>::epsilon())
-                    {
-                        const double FluidWeight = g * rProp[DENSITY_WATER];
-
-                        array_1d<double, 3> NodeCoordinates;
-                        noalias(NodeCoordinates) = rGeom[node].Coordinates();
-                        array_1d<double, 3> NodeVolumeAccelerationUnitVector;
-                        noalias(NodeVolumeAccelerationUnitVector) = NodeVolumeAcceleration / g;
-
-                        const double WaterPressure = rGeom[node].FastGetSolutionStepValue(WATER_PRESSURE);
-                        rGeom[node].SetValue(*element_var, -inner_prod(NodeCoordinates, NodeVolumeAccelerationUnitVector) - PORE_PRESSURE_SIGN_FACTOR * WaterPressure / FluidWeight);
-                    }
-                    else
-                    {
-                        rGeom[node].SetValue(*element_var, 0.0);
-                    }
-                }
-            }
-
-            gid_io.WriteNodalResultsNonHistorical(*element_var, model_part.Nodes(), 0);
     }
 
     int KratosExecute::mainExecution(ModelPart &model_part,
@@ -718,11 +313,11 @@ namespace Kratos
 
             KRATOS_INFO_IF("GeoFlowKernel", this->GetEchoLevel() > 0) << "Nodal Solution Variables Added" << std::endl;
 
-            parseMesh(model_part, meshpath);
+            KratosGeoParser::parseMesh(model_part, meshpath);
 
             KRATOS_INFO_IF("GeoFlowKernel", this->GetEchoLevel() > 0) << "Parsed Mesh" << std::endl;
 
-            parseMaterial(current_model, materialpath);
+            KratosGeoParser::parseMaterial(current_model, materialpath);
 
             KRATOS_INFO_IF("GeoFlowKernel", this->GetEchoLevel() > 0) << "Parsed Material" << std::endl;
 
@@ -734,7 +329,7 @@ namespace Kratos
 
             KRATOS_INFO_IF("GeoFlowKernel", this->GetEchoLevel() > 0) << "Added DoF" << std::endl;
 
-            std::vector<std::shared_ptr<Process>> processes = parseProcess(model_part, projectfile);
+            std::vector<std::shared_ptr<Process>> processes = KratosGeoParser::parseProcess(model_part, projectfile);
 
             KRATOS_INFO_IF("GeoFlowKernel", this->GetEchoLevel() > 0) << "Parsed Process Data" << std::endl;
 
@@ -751,7 +346,7 @@ namespace Kratos
             if (!hasPiping)
             {
                 mainExecution(model_part, processes, p_solving_strategy, 0.0, 1.0, 1);
-                outputGiD(current_model, model_part, projectfile, workingDirectory);
+                KratosGeoOutput::KratosGeoOutput::outputGiD(current_model, model_part, projectfile, workingDirectory);
             }
             else
             {
@@ -836,7 +431,7 @@ namespace Kratos
                         break;
                     }
 
-                    outputGiD(current_model, model_part, projectfile, workingDirectory);
+                    KratosGeoOutput::KratosGeoOutput::outputGiD(current_model, model_part, projectfile, workingDirectory);
 
                     // Update boundary conditions for next search head.
                     if (RiverBoundary->Info() == "ApplyConstantScalarValueProcess")
