+++ applications/GeoMechanicsApplication/tests/test_compare_sellmeijer/HeightAquiferD10L30.gid/criticalHead.json
@@ -1,0 +1,6 @@
+{
+	 "PipeData":	{
+		 "Success": "True",
+		 "CriticalHead": "3.700000"
+	 }
+}
